Added Category Blacklist Feature
Some users might not want to blanket add all torrents to their database

so I've added a category blacklist feature to exclude torrents of

certain categories. The feature is case sensitive i.e myCategory is not

the same as MYCATEGORY. Enhancements could be made to support more

filtering perhaps through the use of wildcards and regular expressions.

Configuration is done in the config.json file as follows:

"blacklisted_categories": [

"Category to blacklist 1",

"Category to blacklist 2"

]